requirements.txt
django==4.2.14
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-redis==5.4.0
redis==5.0.7
drf-spectacular==0.27.2
python-dotenv==1.0.1
dj-database-url==3.0.1
pytest==8.3.2
pytest-django==4.9.0
gunicorn==21.2.0
psycopg2-binary==2.9.9

